[DEV] Add setting surviving time to 14.59 if user has failed on last boss

+++ Assets/Scripts/TonPlay.Client.Roguelike/UI/Screens/Game/Timer/Interfaces/ITimerContext.cs
@@ -6,6 +6,6 @@ namespace TonPlay.Client.Roguelike.UI.Screens.Game.Timer.Interfaces
 {
 	public interface ITimerContext : IScreenContext
 	{
-		IReadOnlyReactiveProperty<float> TimeInSeconds { get; }
+		IReadOnlyReactiveProperty<double> TimeInSeconds { get; }
 	}
 }
